+++ app/src/main/java/com/example/razomua/model/Profile.kt
@@ -2,8 +2,8 @@ package com.example.razomua.model
 import kotlinx.serialization.Serializable
 @Serializable
 data class Profile(
-    val id: Long,
-    val userId: Long,
+    val id: Int,
+    val userId: Int,
     val photoUrl: String?,
     val location: String?
 )
