+++ app/src/main/java/com/example/razomua/model/ProfileInterest.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ProfileInterest(
-    val profileId: Long,
-    val interestId: Long
+    val profileId: Int,
+    val interestId: Int
 )
